refactor(tests): decentralize monolithic unit.spec.ts and add state boundary tests
